# public/api.yaml
info:
  title: Coronatime API
  description: Here goes description
  version: 0.1.9
servers:
  - url: http://localhost:3000/wane/v1
    description: local server
  - url: http://producation.app/api
    description: production server
openapi: 3.0.0
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /login:
    post:
      operationId: login
      summary: Login
      description: Authenticate a user and return a token
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
              required:
                - login
                - password
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      refresh_token:
                        type: string
        '401':
          description: Unauthenticated
  /register:
    post:
      operationId: register
      summary: Register a new user
      description: Register a new user with the provided details
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nom:
                  type: string
                prenom:
                  type: string
                login:
                  type: string
                password:
                  type: string
                password_confirmation:
                  type: string
                photo:
                  type: string
                client_id:
                  type: integer
              required:
                - nom
                - prenom
                - login
                - password
                - password_confirmation
                - client_id
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        nom:
                          type: string
                        prenom:
                          type: string
                        login:
                          type: string
                        photo:
                          type: string
                  message:
                    type: string
                  status:
                    type: string
        '401':
          description: Unauthenticated
  /logout:
    post:
      operationId: logout
      summary: Logout
      description: Logout
      tags:
        - Auth
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthenticated
...